Revert "Merge pull request #67412 from callstack-internal/duplicate-workspace"
This reverts commit 8f4897f001a2f8977480fb8fd9ba9cfb723dcd74, reversing
changes made to 485a092e3c00483f259420b593e4403045abe5b6.
