+++ src/ROUTES.ts
@@ -1876,14 +1876,6 @@ const ROUTES = {
             return getUrlWithBackToParam(`workspaces/${policyID}/per-diem`, backTo);
         },
     },
-    WORKSPACE_DUPLICATE: {
-        route: 'workspace/:policyID/duplicate',
-        getRoute: (policyID: string, backTo?: string) => getUrlWithBackToParam(`workspace/${policyID}/duplicate`, backTo),
-    },
-    WORKSPACE_DUPLICATE_SELECT_FEATURES: {
-        route: 'workspace/:policyID/duplicate/select-features',
-        getRoute: (policyID: string, backTo?: string) => getUrlWithBackToParam(`workspace/${policyID}/duplicate/select-features`, backTo),
-    },
     WORKSPACE_RECEIPT_PARTNERS: {
         route: 'workspaces/:policyID/receipt-partners',
         getRoute: (policyID: string | undefined, backTo?: string) => {
