+++ src/languages/zh-hans.ts
@@ -320,7 +320,6 @@ const translations = {
         count: '计数',
         cancel: '取消',
         dismiss: '忽略',
-        proceed: 'Proceed',
         yes: '是的',
         no: '不',
         ok: '好的',
@@ -3430,14 +3429,12 @@ const translations = {
             customField1: '自定义字段 1',
             customField2: '自定义字段2',
             customFieldHint: '添加适用于该成员所有支出的自定义编码。',
-            reports: '报告',
             reportFields: '报告字段',
             reportTitle: '报告标题',
             reportField: '报告字段',
             taxes: '税款',
             bills: '账单',
             invoices: '发票',
-            perDiem: 'Per diem',
             travel: '旅行',
             members: '成员',
             accounting: '会计',
@@ -3450,7 +3447,6 @@ const translations = {
             testTransactions: '测试交易',
             issueAndManageCards: '发行和管理卡片',
             reconcileCards: '对账卡片',
-            selectAll: '全选',
             selected: () => ({
                 one: '1 已选择',
                 other: (count: number) => `已选择${count}个`,
@@ -3464,8 +3460,6 @@ const translations = {
             memberNotFound: '未找到成员。要邀请新成员加入工作区，请使用上面的邀请按钮。',
             notAuthorized: `您无权访问此页面。如果您正在尝试加入此工作区，请请求工作区所有者将您添加为成员。还有其他问题？请联系${CONST.EMAIL.CONCIERGE}。`,
             goToWorkspace: '前往工作区',
-            duplicateWorkspace: '重复工作区',
-            duplicateWorkspacePrefix: '复制',
             goToWorkspaces: '前往工作区',
             clearFilter: '清除筛选器',
             workspaceName: '工作区名称',
@@ -4832,18 +4826,6 @@ const translations = {
             importedFromAccountingSoftware: '以下税费是从您的',
             taxCode: '税码',
             updateTaxCodeFailureMessage: '更新税码时发生错误，请重试',
-        },
-        duplicateWorkspace: {
-            title: '命名您的新工作区',
-            selectFeatures: '选择要复制的功能',
-            whichFeatures: '您想要将哪些功能复制到您的新工作区？',
-            confirmDuplicate: '\n\n您想继续吗？',
-            categories: '类别和您的自动分类规则',
-            reimbursementAccount: '报销账户',
-            delayedSubmission: '延迟提交',
-            welcomeNote: '请开始使用我的新工作区',
-            confirmTitle: ({newWorkspaceName, totalMembers}: {newWorkspaceName?: string; totalMembers?: number}) =>
-                `您即将创建并与原始工作区中的 ${totalMembers ?? 0} 名成员共享 ${newWorkspaceName ?? ''}。`,
         },
         emptyWorkspace: {
             title: '您没有任何工作区',
