+++ src/libs/Navigation/AppNavigator/Navigators/RightModalNavigator.tsx
@@ -149,10 +149,6 @@ function RightModalNavigator({navigation, route}: RightModalNavigatorProps) {
                             component={ModalStackNavigators.WorkspaceConfirmationModalStackNavigator}
                         />
                         <Stack.Screen
-                            name={SCREENS.RIGHT_MODAL.WORKSPACE_DUPLICATE}
-                            component={ModalStackNavigators.WorkspaceDuplicateModalStackNavigator}
-                        />
-                        <Stack.Screen
                             name={SCREENS.RIGHT_MODAL.NEW_TASK}
                             component={ModalStackNavigators.NewTaskModalStackNavigator}
                         />
