--- src/types/form/WorkspaceDuplicateForm.tsx
@@ -1,18 +1,0 @@
-import type {ValueOf} from 'type-fest';
-import type Form from './Form';
-
-const INPUT_IDS = {
-    NAME: 'name',
-} as const;
-
-type InputID = ValueOf<typeof INPUT_IDS>;
-
-type WorkspaceDuplicateForm = Form<
-    InputID,
-    {
-        [INPUT_IDS.NAME]: string;
-    }
->;
-
-export type {WorkspaceDuplicateForm};
-export default INPUT_IDS;
