+++ src/types/form/index.ts
@@ -34,7 +34,6 @@ export type {ReportVirtualCardFraudForm} from './ReportVirtualCardFraudForm';
 export type {DebugReportForm} from './DebugReportForm';
 export type {DebugReportActionForm} from './DebugReportActionForm';
 export type {DebugTransactionForm} from './DebugTransactionForm';
-export type {WorkspaceDuplicateForm} from './WorkspaceDuplicateForm';
 export type {DebugTransactionViolationForm} from './DebugTransactionViolationForm';
 export type {RequestPhysicalCardForm} from './RequestPhysicalCardForm';
 export type {RoomNameForm} from './RoomNameForm';
